refactor(organism_test.go): remove paralleltest and add comment
The searchLimit variable was of type int64, but was being assigned
an int value. This commit fixes the type mismatch. The tests in
organism_test.go were not safe for parallel execution because they
depend on shared data. The paralleltest directive was removed and a
comment was added to indicate that the tests depend on shared data.
The annListQ variable was not used and was removed.

+++ internal/app/service/read_service.go
@@ -81,7 +81,7 @@ func (srv *AnnotationService) ListAnnotationGroups(
 		return nil, aphgrpc.HandleInvalidParamError(ctx, err)
 	}
 	// default value of limit
-	searchLimit := int64(LIMIT)
+	searchLimit := LIMIT
 	if rgp.Limit > 0 {
 		searchLimit = rgp.Limit
 	}
